+++ src/main/java/com/example/autoservice/dto/response/OrderResponseDto.java
@@ -1,6 +1,8 @@
 package com.example.autoservice.dto.response;
 
 import com.example.autoservice.model.Status;
+
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.List;
 import lombok.Data;
@@ -14,6 +16,6 @@ public class OrderResponseDto {
     private List<Long> carServicesId;
     private List<Long> productsId;
     private Status status;
-    private double price;
+    private BigDecimal price;
     private LocalDate finishDate;
 }
